requirements_clean.txt
# Core Flask Framework
Flask==2.3.3
Werkzeug==2.3.7

# Database
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.11
alembic

# Authentication & Security
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Session==0.5.0
authlib
google-auth-oauthlib
pyotp==2.9.0

# Environment & Configuration
python-dotenv==1.0.0
pytz==2023.3

# HTTP Requests
requests==2.31.0

# Trading APIs
dhanhq
kiteconnect
SmartApi

# Payment Gateway
razorpay==1.3.0

# Utilities
beautifulsoup4
lxml
pandas>=2.1.0
numpy

# Template Engine (included with Flask but explicit for clarity)
Jinja2
MarkupSafe

# Development Tools (optional)
flask-cors